create oneTrustAssessmentToCsvRecord helper
